Change embedding method return type to List<double>
Updated the return type of the `GenerateEmbeddingAsync` method in both the `IEmbeddingService` interface and the `EmbeddingService` class from `List<float>` to `List<double>`. This change enhances precision for the embeddings generated by the service.

+++ AIKnowledgeAssistant.Api/Interfaces/IEmbeddingService.cs
@@ -2,7 +2,7 @@
 {
     public interface IEmbeddingService
     {
-        Task<List<float>> GenerateEmbeddingAsync(string text);
+        Task<List<double>> GenerateEmbeddingAsync(string text);
 
     }
 }
